requirements.txt
allure-pytest==2.13.1
allure-python-commons==2.13.1
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
exceptiongroup==1.1.1
greenlet==2.0.1
idna==3.4
iniconfig==2.0.0
packaging==23.0
playwright==1.31.1
pluggy==1.0.0
pyee==9.0.4
pytest==7.2.2
python-slugify==6.1.2
requests==2.28.2
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
